Src/BackEnd/Search/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0

# Testing and development
pytest==7.4.2
pytest-cov==4.1.0

# Utility packages
python-dotenv==1.0.0  # For environment variable management
requests==2.31.0      # For HTTP requests
tqdm==4.66.1         # For progress bars


# Web framework
pydantic==2.4.2
python-multipart==0.0.6

# Database
pymongo==4.6.0

# ML Libraries - CPU optimized versions
torch==2.1.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
numpy==1.24.3
scikit-learn==1.3.0
rank-bm25==0.2.2

# Transformers - minimal installation
transformers==4.35.0
tokenizers==0.14.1
sentence-transformers==2.2.2

# Hugging Face
huggingface-hub==0.16.4

# OpenAI
openai==1.95.1

# Additional CPU optimizations
# Exclude heavy optional dependencies
--extra-index-url https://download.pytorch.org/whl/cpu
